requirements.txt
aiogram==3.25.0
aiosqlite==0.22.1
aiohttp==3.13.3
openai==2.20.0
python-dotenv==1.2.1
